Implement offline favourite posts and files viewing
